still refactoring configs
