+++ cll/__init__.py
@@ -1,12 +1,1 @@
-OPENAI_DEFAULT_PARAMS = {
-    "frequency_penalty": 0,
-    "logprobs": 1,
-    "max_tokens": 200,
-    "model": "gpt-3.5-turbo",
-    "n": 1,
-    "presence_penalty": 0,
-    "stop": ["I'm sorry", "As an AI"],
-    "temperature": 0.9,
-    "top_p": 1,
-    "stream": True,
-}
+# Init
