+++ cll/context.py
@@ -1,0 +1,36 @@
+# @staticmethod
+# def context(_, command_params, tree):
+#     if command_params and command_params[0] == "help":
+#         click.echo(
+#             "modify the context. context can be added to nodes but are not part of the main path\n"
+#             "\t\t\tsubcommands are clear, list, remove or add"
+#         )
+#         return
+#
+#     if command_params[0] == "clear":
+#         for node in tree.index.path:
+#             if node.node_info.get("context"):
+#                 del node.node_info["context"]
+#
+#     if command_params[0] == "list":
+#         docs = [(tree.index.get_context(node), node.index) for node in tree.index.path]
+#         docs = [(doc, index) for doc, index in docs if doc is not None]
+#         for doc, index in docs:
+#             doc_text = doc.text.replace("\n", " ")[: tree.termwidth]
+#             click.echo(f"{index}: {doc_text}")
+#         return
+#
+#     if command_params[0] == "remove":
+#         for index in command_params[1:]:
+#             tree.index.delete_context(int(index))
+#
+#     if command_params[0] == "add":
+#         # If no context is given, just add the last node as context
+#         if len(command_params) == 1:
+#             new_context = tree.index.path[-1].text
+#         else:
+#             new_context = " ".join(command_params[1:])
+#
+#         tree.index.add_context(new_context, tree.index.path[-1])
+#
+#     tree.save()
